Implemented wall following and distance measuring

+++ include/zones.h
@@ -20,7 +20,26 @@
  */
 short mazeSolving(short);
 
-void wallFollowing(int);
+/**
+ * Procedure to operate bot in section B-C and C-D. The zone includes:
+ *  - Wall following
+ *  - Wall switching
+ *  - Front wall detection (Turn on opposite to primary side)
+ *  - Turn on primary side
+ * Wall is followed while maintaing an average distance, which acts as the center line between min and max distance.
+ *  
+ * @param primary Initial side on which wall is present.
+ */
+void wallFollowing(byte);
+
+/**
+ * Procedure to operate bot in section D-E. The zone includes:
+ *  - Straight line following
+ *  - Node detection
+ *  - Measuring distance between two nodes
+ * When the first node is detected, the encoder is initialised. It is stopped when the second node is detected.
+ * The encoding is done using a secondary controller which acts as slave. Distance is requested once second node is detected.
+ */
 void distanceMeasuring();
 
 #endif
